pyproject.toml
[tool.poetry]
name = "bytecover"
version = "0.1.0"
description = ""
authors = ["petvav <peter@orfium.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
ffmpeg-python = "^0.2.0"
pandas = "^2.0.0"
numpy = "~1.23"
pyyaml = "^6.0"
scikit-learn = "^1.2.2"
protobuf = "^3.20"
wandb = "^0.14.0"
optuna = "^3.1.0"
torch = "^2.0.0"
torchaudio = "^2.0.1"
tqdm = "^4.65.0"
nnaudio = "^0.3.2"
torchvision = "^0.15.1"
jsonlines = "^3.1.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.2.2"
pylint = "^2.17.2"
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
